all functions working

+++ src/user/UserRouter.js
@@ -12,7 +12,6 @@ UserRouter.route('/account').get(requireAuth, (req, res, next) => {
       res.json(profile);
     })
     .catch(err => {
-      console.log('Account error', err);
       next(err);
     });
 });
@@ -26,28 +25,26 @@ UserRouter.route('/favorites')
         res.json(profile);
       })
       .catch(err => {
-        console.log('Account error', err);
         next(err);
       });
   })
   .post(requireAuth, jsonBodyParser, (req, res, next) => {
-    console.log('WE ARE CONSOLE LOGGING THE BODY HERE');
-    console.log(req.body);
     const { venue_id } = req.body;
     const newFavorite = { venue_id };
 
-    //check for validation here
-
-    //add user id from authtoken
-
     newFavorite.user_id = req.user.id;
+
+    if(!newFavorite.user_id) {
+      return res.status(401).json( {
+        error: `Unauthorized`
+      });
+    }
+
     UserService.addFavorite(req.app.get('db'), newFavorite)
       .then(favorite => {
-        console.log('eyyyyyyy!');
         res.json(favorite);
       })
       .catch(err => {
-        console.log('Favorties error', err);
         next(err);
       });
   })
@@ -55,16 +52,34 @@ UserRouter.route('/favorites')
     const { venue_id } = req.body;
     const delFav = { venue_id };
 
-    //MACHEN SOME VALIDATION HERE!
-
-    //SET THE user...
-
     delFav.user_id = req.user.id;
+
+    if (!delFav.user_id) {
+      return res.status(401).json({
+        error: `Unauthorized`
+      });
+    }
+
     UserService.deleteFavorite(req.app.get('db'), delFav)
       .then(numRowsAffected => {
         res.status(204).end();
       })
       .catch(next);
   });
+
+
+  UserRouter.route('/userReviews')
+  .all(requireAuth)
+  .get(requireAuth, (req, res, next) => {
+    UserService.getUserReviews(req.app.get('db'), req.user.id)
+      .then(profile => {
+        res.json(profile);
+      })
+      .catch(err => {
+        console.log('Account error', err);
+        next(err);
+      });
+  });
+
 
   module.exports = UserRouter;
